Part1: add MatchImagePairs and binary code determination
Summary:
The goal of this diff is to implement two-shot detection for isometric pattern with an idea of:

Add an additional MatchImagePair() function to PatternMatcherIsometric, and when we use this pattern matcher for display calibration, we directly create a unique ptr of PatternMatcherIsometric instead of IPatternMatcher.

In this step we add an additional MatchImagePair() function to PatternMatcherIsometric and binary code determination for two-shot images input.

Reviewed By: YLouWashU

Differential Revision: D29215200

fbshipit-source-id: 4ddd34622d6cadc14fc5187ef4962f3fb9a5f4d7