Part2: find a bug in DotDetection.h
Summary:
The goal of this diff is to implement two-shot detection for isometric pattern with an idea of:

Add an additional MatchImagePair() function to PatternMatcherIsometric, and when we use this pattern matcher for display calibration, we directly create a unique ptr of PatternMatcherIsometric instead of IPatternMatcher.

In this step we found a bug in DotDetection.h and fixed it with some hacky solution. We will fix the bug later.

Differential Revision: D29215623

fbshipit-source-id: 50b18f602c792ab7f7fec47aaacf9a38bfdde730