support numberBlock and goodPoseInlierRatio input + fix a small bug in code
Summary:
We support API for `numberBlock` and `goodPoseInlierRatio`.

`numberBlock`: how many blocks in row/column direction will be picked. So final number of blocks will be `numberBlock * numberBlock`, and we calculate pose from each block and select good poses within them.

`goodPoseInlierRatio`: the threshold of selecting good pose. The good selected pose should have the number of inliers dots larger than `goodPoseInlierRatio ` * total number of dots.

We also fixed a bug about updating `bfsProcessSeq_`. Previously, `bfsProcessSeq_` is updated in the for loop (L647) when `poseIdx >=1`. When `poseIdx = 0`, namely `transferDotsGroup.size() == 1`, `bfsProcessSeq_` will not be updated so we have to update it ahead.

Differential Revision: D29855302

fbshipit-source-id: b2d2ce8927b7b45fb50ea8d5bf4a114517e4e4af